+++ src/main/java/com/github/zillow/web/dto/ListingDTO.java
@@ -2,10 +2,7 @@ package com.github.zillow.web.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import jakarta.persistence.Column;
 import lombok.*;
-
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -19,9 +16,9 @@ public class ListingDTO {
     private String city;
     private String state;
     private String zipcode;
-    private BigDecimal Latitude;
-    private BigDecimal Longitude;
-    private Double price;
+    private double Latitude;
+    private double Longitude;
+    private double price;
     private Integer bedrooms;
     private Integer bathrooms;
     private Integer sqft;
